docugen-ai/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
email-validator==2.2.0
openai==1.59.7
python-docx==1.1.2
docxtpl==0.19.1
reportlab==4.2.5
mammoth==1.9.0
pypandoc==1.14
Pillow==11.1.0
docx2pdf==0.1.8
